+++ clang/lib/CMakeLists.txt
@@ -10,6 +10,7 @@ add_subdirectory(Sema)
 add_subdirectory(CodeGen)
 add_subdirectory(Analysis)
 add_subdirectory(Edit)
+add_subdirectory(ExtractAPI)
 add_subdirectory(Rewrite)
 if(CLANG_ENABLE_ARCMT)
   add_subdirectory(ARCMigrate)
@@ -24,7 +25,6 @@ add_subdirectory(Index)
 add_subdirectory(IndexDataStore)
 add_subdirectory(IndexSerialization)
 add_subdirectory(StaticAnalyzer)
-add_subdirectory(SymbolGraph)
 add_subdirectory(Format)
 add_subdirectory(Testing)
 add_subdirectory(Interpreter)
