+++ clang/test/ExtractAPI/global_record.c
